[fix]mypage api.js 생성 및 분리, 엔드포인트 env 숨김
