+++ src/pages/myPage/mypage/useScrapStore.js
@@ -9,3 +9,4 @@ const useScrapStore = create(set => ({
 }));
 
 export default useScrapStore;
+
